dna_alignment_manager/requirements.txt
biopython==1.74
numpy==1.17.3
Django==2.2.6
